Initial setup of routing and page structure
