+++ package.json
@@ -14,7 +14,8 @@
     "bootstrap": "^5.1.1",
     "core-js": "^3.6.5",
     "d3": "^7.0.1",
-    "vue": "^3.0.0"
+    "vue": "^3.0.0",
+    "vue-router": "4"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
