+++ src/components/About.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>
+    <h1>About</h1>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "About"
+}
+</script>
+
+<style scoped>
+
+</style>
