+++ src/components/Introduction.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>
+    <h1>Introduction</h1>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "Introduction"
+}
+</script>
+
+<style scoped>
+
+</style>
